ezyZip/requirements.txt
google-generativeai
reportlab
fpdf